requirements.txt
absl-py==0.6.1
astor==0.7.1
biopython==1.72
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
panda==0.3.1
protobuf==3.6.1
pydot==1.4.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
requests==2.20.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
